requirements.txt
# Model Behavior Stability & Consistency Analyzer
# Minimal dependencies for evaluation system

# Core dependencies
pyyaml>=6.0
numpy>=1.21.0

# Optional: For real model inference (if transformers is available)
# Uncomment the following line if you want to use actual transformer models:
# transformers>=4.20.0
# torch>=1.12.0

# Note: The system will work with mock inference if transformers/torch are not installed

